pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nocode"
version = "0.1.0"
description = "NOCODE — autonomous, terminal-native coding agent."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "pythongiant" }]
dependencies = [
    "openai>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.scripts]
nocode = "nocode.main:main"

[tool.setuptools.packages.find]
include = ["nocode*"]

[tool.setuptools.package-data]
nocode = ["defaults/*"]
